+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for global rules.
 - Fix origin field of tool calls.
+- Allow chat communication with no workspace opened.
 
 ## 0.3.1